requirements.txt
django 3.2.7
telebot
requests
os
djangorestframework